ToDo Dockerize, tests and validation fields on backend

+++ products/models.py
@@ -3,8 +3,8 @@ from django.db import models
 # Create your models here.
 class Products(models.Model):
     name = models.CharField(max_length=255)
-    price = models.DecimalField(max_digits=999999, decimal_places=2)
-    cost = models.DecimalField(max_digits=999999, decimal_places=2)
+    price = models.FloatField()
+    cost = models.FloatField()
     company = models.ForeignKey("companies.company", on_delete=models.CASCADE)
 
     def __str__(self):
